+++ anikanov.vladislav/T1/shape.h
@@ -13,11 +13,11 @@ public:
 
   virtual void move(PointT) = 0;
 
-  virtual void move(double x, double y) = 0;
+  virtual void move(float x, float y) = 0;
 
   virtual PointT getCPoint() const = 0;
 
-  virtual void scale(double) = 0;
+  virtual void scale(float) = 0;
 };
 
 #endif
